moved custom settings button story example into Toolbar story
